move localization folder
