+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:book_adapter/app.dart';
+import 'package:book_adapter/src/app.dart';
 import 'package:book_adapter/src/features/in_app_update/update.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
